Move Embed into utilites.
