+++ resources/assets/components/utilities/Embed/embed.scss
@@ -1,4 +1,4 @@
-@import '../../scss/next-toolbox';
+@import '../../../scss/next-toolbox';
 
 .embed {
   display: flex;
